Centralize scoped introspection contracts
