+++ graphile/graphile-settings/package.json
@@ -62,7 +62,6 @@
     "graphile-presigned-url-plugin": "workspace:^",
     "graphile-realtime-subscriptions": "workspace:^",
     "graphile-search": "workspace:^",
-    "graphile-scoped-introspection": "workspace:^",
     "graphile-upload-plugin": "workspace:^",
     "graphile-utils": "5.0.3",
     "graphql": "16.13.0",
